# improve the performance of reverse_complement